Scope viewModel on ArticleScreen to ensure we don't re-use a ViewModel with the previous postId.
This is a partial workaround for https://issuetracker.google.com/issues/165642391

Currently, viewModel() in compose is scoped ViewModel to the lifecycle of the Activity or Fragment that
calls the composable. This causes the the ViewModel to be re-used when created multiple times in composition.

On a detail screen where we pass an id parameter to the ViewModel constructor, the wrong data would be loaded
when leaving and entering the screen.

This commit introduces a partial solution by adding a way to scope ViewModels to the lifecycle of the composable
that called them. On review, we decided that this code should not be shipped as the example implementation does
not retain instances on rotation and should not be used in production.

Warning: It is not recommended to use ViewModelLifecycleScope as it does not retain instances on rotation.

Follow https://issuetracker.google.com/issues/165642391 for more updates as these APIs are worked on.

Various code cleanup

- prefer observe* name for Flow returning functions in Repository
- Make Set<T>.addOrRemove an internal extension function and use it throughout
- Deprecate sample implementation of ViewModelLifecycleScope as it is intended to be replaced with a public API
- Add missing comments to ArticleViewModel.kt
- Corrected comments about viewModel scope to say Activity instead of Application
- Added MapExtensions.kt that contains an internal extension addOrRemove
- Swapped to use the liveData builder for ArticleViewModel.post