[Jetsurvey] Fix for radio button accessibility
This moves the selectable Modifier to the Surface of the Answer
composable. Since the SingleChoiceQuestion composable adds the
selectableGroup Modifier, the Answers will now act as a group of
mutually exclusive radio buttons.

Note that the clip Modifier has to be added for the ripple. In the
future, we can refactor this to use the selectable or clickable versions
of Surface, which handle this already.