[Jetsurvey] Simplification of survey screens
This eliminates the survey repository and related data models in favor
of simplifying the code base. Each of the different question types has a
generic composable such as MultipleChoiceQuestion, which takes all of
the relevant strings/resources to generate the UI. The actual survey
data is baked into a specific composable in Survey.kt, such as
FreeTimeQuestion, instead of coming from a data model.

The SurveyViewModel is now much more explicit with methods for each
question such as onFreeTimeResponse. In a production app, you'd likely
want this to be more flexible to handle server-side questions, but this
code pattern makes the sample easier to skim. Every response is
maintained in SurveyViewModel as a separate state. This increases
verbosity in favor of better readability.

This also appears to fix the issue of photos not working on API 21,
though I'm not sure what specific change fixed it. I verified on a Nexus
4 running Android 5.1.1.

Note: This does not animate between the final survey question and the
survey results. When the work to move the app to Compose Navigation is
done, the results will be a separate destination that is navigated to
and animated that way.

This fixes #1054.